Refactor routing structure: replace router.js with AppRouter.js, update navigation paths, and enhance report creation logic in CreateReport component.
