+++ src/features/users/Dashboard.js
@@ -17,7 +17,7 @@ import {
 import { useTheme } from "@mui/material/styles";
 import { useLoaderData, useNavigate } from "react-router";
 import { userService } from "./user.service";
-import { adminService } from "../../services/admin.service";
+import { reportService } from "../reports/report.service";
 
 export async function dashboardLoader() {
   const user = await userService.refreshToken();
@@ -25,7 +25,7 @@ export async function dashboardLoader() {
     throw new Response("dashboardLoader user problem", { status: 500 });
   }
   //   const user = userService.userValue; // Get the current user
-  const records = await adminService.getAllById({ clientId: user.clientId });
+  const records = await reportService.getAllById({ clientId: user.clientId });
   if (!records) {
     throw new Response("dashboardLoader records problem", { status: 500 });
   }
@@ -61,8 +61,8 @@ export default function Dashboard() {
 
   function createReport(report) {
     // Logic to create a new report
-    navigate("/report-create", {
-      state: { reportName: report.name },
+    navigate("/create-report", {
+      state: { reportName: report.name, reportList },
     });
   }
